cost-auditor/backend/collectors/gcp_collector.py
"""
Google Cloud Platform Cost Collector
"""
import os
from google.cloud import billing_v1
from google.cloud import monitoring_v3
from datetime import datetime, timedelta
from typing import Dict, List

class GCPCollector:
    def __init__(self):
        self.project_id = os.getenv('GCP_PROJECT_ID', 'opencellcw-k8s')
        self.billing_account = os.getenv('GCP_BILLING_ACCOUNT')

    def get_usage(self, start_date: datetime = None, end_date: datetime = None) -> Dict:
        """Get GCP usage and costs"""
        if not start_date:
            start_date = datetime.now() - timedelta(days=30)
        if not end_date:
            end_date = datetime.now()

        try:
            costs = self._get_billing_data(start_date, end_date)
            resources = self._get_resource_usage()

            return {
                'service': 'gcp',
                'project_id': self.project_id,
                'period': {
                    'start': start_date.isoformat(),
                    'end': end_date.isoformat()
                },
                'costs': costs,
                'resources': resources
            }

        except Exception as e:
            return {'error': str(e)}

    def _get_billing_data(self, start_date: datetime, end_date: datetime) -> Dict:
        """
        Get billing data from Cloud Billing API

        Requires billing.accounts.getCostTable permission
        """
        try:
            # Initialize billing client
            client = billing_v1.CloudBillingClient()

            # Query billing data
            # Note: This is a simplified example. Actual implementation
            # would use BigQuery billing export or Cloud Billing API

            return {
                'total': 0.0,
                'by_service': {
                    'kubernetes_engine': 0.0,
                    'compute_engine': 0.0,
                    'cloud_storage': 0.0,
                    'cloud_build': 0.0,
                    'artifact_registry': 0.0,
                    'networking': 0.0
                },
                'currency': 'USD'
            }

        except Exception as e:
            return {'error': str(e)}

    def _get_resource_usage(self) -> Dict:
        """Get detailed resource usage"""
        try:
            # GKE cluster details
            cluster_usage = self._get_gke_usage()

            # Storage usage
            storage_usage = self._get_storage_usage()

            return {
                'gke': cluster_usage,
                'storage': storage_usage
            }

        except Exception as e:
            return {'error': str(e)}

    def _get_gke_usage(self) -> Dict:
        """Get GKE cluster usage"""
        # Would use GKE API to get:
        # - Node count and types
        # - Pod count
        # - Resource requests/limits
        # - Actual resource usage

        return {
            'cluster_name': 'ulf-cluster',
            'zone': 'us-central1-a',
            'node_count': 3,
            'node_type': 'e2-medium',
            'total_vcpus': 6,
            'total_memory_gb': 12
        }

    def _get_storage_usage(self) -> Dict:
        """Get storage usage"""
        # Would query Cloud Storage and Persistent Disk usage
        return {
            'persistent_disks': {
                'total_gb': 15,
                'cost_per_gb_month': 0.04
            },
            'cloud_storage': {
                'total_gb': 0,
                'cost_per_gb_month': 0.02
            }
        }

    def get_current_month_cost(self) -> float:
        """Get current month cost"""
        start_of_month = datetime.now().replace(day=1, hour=0, minute=0, second=0)
        data = self.get_usage(start_date=start_of_month)
        return data.get('costs', {}).get('total', 0.0)

    def get_cost_breakdown(self) -> Dict:
        """Get detailed cost breakdown by service"""
        data = self.get_usage()
        return data.get('costs', {}).get('by_service', {})
