cost-auditor/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
requests==2.31.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Google Cloud (for GCP billing)
google-cloud-billing==1.12.0
google-cloud-monitoring==2.18.0

# Scheduling
apscheduler==3.10.4

# Data visualization (for future dashboard)
pandas==2.1.4
plotly==5.18.0
